refactor: 🎨 improve structure
