+++ lib/parking_manager/presentation/get_history/get_history_page.dart
@@ -1,12 +1,11 @@
 import 'package:date_time_picker/date_time_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:parking_manager/parking_manager/domain/entities/parking_entity.dart';
 import 'package:parking_manager/parking_manager/presentation/get_history/bloc/get_history_bloc.dart';
-import 'package:parking_manager/shared/themes/app_colors.dart';
+import 'package:parking_manager/parking_manager/presentation/get_history/widgets/generic_history_list_view_widget.dart';
 import 'package:parking_manager/shared/themes/app_text_styles.dart';
-import 'package:parking_manager/shared/utils/extensions/currency_extension.dart';
-import 'package:parking_manager/shared/utils/extensions/date_extension.dart';
+import 'package:parking_manager/shared/widgets/error.dart';
+import 'package:parking_manager/shared/widgets/loading.dart';
 
 class GetHistoryPage extends StatelessWidget {
   const GetHistoryPage({
@@ -16,8 +15,8 @@ class GetHistoryPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final historyBloc = BlocProvider.of<GetHistoryBloc>(context);
-    TextEditingController dateController = TextEditingController(
-        text: DateTime.now().toIso8601String().split('T')[0]);
+    TextEditingController dateController =
+        TextEditingController(text: formatDateToday());
 
     late Widget finalView;
 
@@ -32,9 +31,8 @@ class GetHistoryPage extends StatelessWidget {
             padding: const EdgeInsets.symmetric(horizontal: 8),
             child: IconButton(
               icon: const Icon(Icons.calendar_month),
-              onPressed: () {
-                showDataPicker(context, historyBloc, dateController);
-              },
+              onPressed: () =>
+                  showDataPicker(context, historyBloc, dateController),
             ),
           ),
         ],
@@ -42,30 +40,20 @@ class GetHistoryPage extends StatelessWidget {
       body: BlocBuilder<GetHistoryBloc, GetHistoryState>(
         builder: (context, state) {
           state.when(
-            initial: () => finalView = HistoryListView(
-              parkingHistoryList: const [],
-              dateController: dateController,
-            ),
-            loading: () => finalView = const Center(
-              child: CircularProgressIndicator(
-                color: AppColors.primary,
-              ),
-            ),
-            loaded: (parking) => finalView = HistoryListView(
-              parkingHistoryList: parking,
-              dateController: dateController,
-            ),
-            error: () => finalView = Center(
-                child: Text(
-              'error',
-              style: AppTextStyles.bold24black(),
-            )),
+            initial: () => finalView = GenericHistoryListViewWidget(
+                parkingHistoryList: const [], dateController: dateController),
+            loading: () => finalView = loadingState,
+            loaded: (parking) => finalView = GenericHistoryListViewWidget(
+                parkingHistoryList: parking, dateController: dateController),
+            error: () => finalView = errorState,
           );
           return finalView;
         },
       ),
     );
   }
+
+  String formatDateToday() => DateTime.now().toIso8601String().split('T')[0];
 
   void showDataPicker(BuildContext context, GetHistoryBloc historyBloc,
       TextEditingController dateController) {
@@ -106,128 +94,6 @@ class GetHistoryPage extends StatelessWidget {
           ],
         );
       },
-    );
-  }
-}
-
-class HistoryListView extends StatelessWidget {
-  final List<ParkingEntity> parkingHistoryList;
-  final TextEditingController dateController;
-  const HistoryListView({
-    super.key,
-    required this.parkingHistoryList,
-    required this.dateController,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.all(16),
-          child: SizedBox(
-            width: MediaQuery.of(context).size.width,
-            child: Card(
-              color: AppColors.greyLight,
-              child: Padding(
-                padding: const EdgeInsets.all(4),
-                child: Text(
-                  'Balance of: ${dateController.text.formatDate()}',
-                  style: AppTextStyles.bold16white(),
-                  textAlign: TextAlign.center,
-                ),
-              ),
-            ),
-          ),
-        ),
-        Expanded(
-          child: ListView.builder(
-            itemCount: parkingHistoryList.length,
-            itemBuilder: (context, index) =>
-                ParkingHistoryItem(parkingEntity: parkingHistoryList[index]),
-          ),
-        ),
-        Padding(
-          padding: const EdgeInsets.all(16),
-          child: SizedBox(
-            width: MediaQuery.of(context).size.width,
-            child: Card(
-              color: AppColors.greyLight,
-              child: Padding(
-                padding: const EdgeInsets.all(16),
-                child: showTotalCost(parkingHistoryList),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Text showTotalCost(List<ParkingEntity> parkingHistoryList) {
-    double totalCost = parkingHistoryList.fold(0.0,
-        (sum, parkintEntity) => sum + ((parkintEntity.parkingCost) ?? 0.0));
-
-    return Text(
-      'The total value of the day was:\n${totalCost.reaisLabel()}',
-      style: AppTextStyles.bold16white(),
-      textAlign: TextAlign.center,
-    );
-  }
-}
-
-class ParkingHistoryItem extends StatelessWidget {
-  final ParkingEntity parkingEntity;
-  const ParkingHistoryItem({
-    super.key,
-    required this.parkingEntity,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16),
-      child: Card(
-        color: AppColors.primaryLight,
-        child: Padding(
-          padding: const EdgeInsets.all(16),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                'Plate: ${parkingEntity.plate}',
-                style: AppTextStyles.bold16white(),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                'Vacancy: ${parkingEntity.vacancy + 1}',
-                style: AppTextStyles.bold16white(),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                'Checkin: ${parkingEntity.checkinTime.formatDateAndHour()}',
-                style: AppTextStyles.bold16white(),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                'Checkout: ${parkingEntity.checkoutTime.formatDateAndHour()}',
-                style: AppTextStyles.bold16white(),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(vertical: 8),
-                child: Divider(
-                  thickness: 2,
-                  color: AppColors.primaryLight,
-                ),
-              ),
-              Text(
-                'Value paid: ${parkingEntity.parkingCost?.reaisLabel()}',
-                style: AppTextStyles.bold16white(),
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
