+++ lib/shared/widgets/error.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'package:parking_manager/shared/themes/app_text_styles.dart';
+
+final errorState = Center(
+  child: Text(
+    'error',
+    style: AppTextStyles.bold24black(),
+  ),
+);
