+++ lib/shared/widgets/loading.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+import 'package:parking_manager/shared/themes/app_colors.dart';
+
+const loadingState = Center(
+  child: CircularProgressIndicator(
+    color: AppColors.primary,
+  ),
+);
